otr/pyproject.toml
[project]
name = "otr"
version = "0.1.0"
description = "OTR"
authors = [
    {name = "Jaganadh Gopinadhan",email = "jaganadhg@users.noreply.github.com"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.9,<=3.13.3"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.78.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
